Add the possibility to use qt stylesheet closes #11
QFrane widgets can now be enhanced using qt stylesheets.
Previously all widgets where QWidgets and where charged with
mgear's widgets but this was a little bit limiting as your needed
to use two files and where limited in terms of widget colors.

The only problem for now is that you can only use cubes and circles